Add COM4J.wrapSta for wrapping STA com objects
Objects wrapped in this way can only be called from the same thread they
were created by (and wrapped in).

When using these wrapped objects calls to the COM object are done in the
same thread instead of using a second thread. This has the disadvantage
that these objects can't be passed around and used from anywhere, but
where a COM object has to be called from a specific thread it means they
can be used (mostly only a problem when the Java code is running in the
same process as an in-process COM server).

Fixes #68
