+++ runtime/src/main/java/com4j/Wrapper.java
@@ -19,8 +19,8 @@ import java.util.WeakHashMap;
  *
  * <p>
  * Every wrapper owns {@link NativePointerPhantomReference} to itself. We'll have this reference queued
- * to {@link ComThread#collectableObjects} when GC determines that the object is no longer needed,
- * or we'll explicitly enqueue it when {@link #dispose()} is called. {@link ComThread} will periodically
+ * to {@link ComThread#getCollectableObjects} when GC determines that the object is no longer needed,
+ * or we'll explicitly enqueue it when {@link #dispose()} is called. {@link ComThreadMulti} will periodically
  * wake up and go through the queue to release interface pointers.
  *
  * @author Kohsuke Kawaguchi (kk@kohsuke.org)
@@ -79,12 +79,18 @@ final class Wrapper implements InvocationHandler, Com4jObject {
      * Wraps a new COM object. The pointer needs to be addRefed by the caller if needed.
      */
     private Wrapper(long ptr) {
-        if(ptr==0)   throw new IllegalArgumentException();
-        assert ComThread.isComThread();
+        if(ptr==0)
+            throw new IllegalArgumentException();
+
+        ComThread thread = Task.getComThread();
+        if (null == thread)
+            thread = ComThreadMulti.get();
+        assert thread.isCurrentThread();
 
         this.ptr = ptr;
-        thread = ComThread.get();
-        ref = new NativePointerPhantomReference(this, thread.collectableObjects, ptr);
+        this.thread = thread;
+
+        ref = new NativePointerPhantomReference(this, thread.getCollectableObjects(), ptr);
         thread.addLiveObject(this);
     }
